cjoy: finish targeting
