+++ ui/app/pages/flags/api.tsx
@@ -99,55 +99,3 @@ export const fetchVariations = async ({
     const { data } = await axios.get(`/flags/${workspaceKey}/${projectKey}/${flagKey}/variations`)
     return data
 }
-
-export type Operator = 'equals' | 'contains' | 'greaterThan' | 'greaterThanOrEqual' | 'contains' | 'regex'
-
-export type TargetingRuleRequest = {
-    key: string
-    name: string
-    description: string
-    tags: string[]
-    type: string
-    traitKey?: string
-    traitValue?: string
-    operator: Operator
-    ruleVariations: {
-        variationKey: string
-        weight: number
-    }[]
-    segmentKey?: string
-}
-
-export type TargetingRuleResponse = {
-    key: string | null | undefined
-    type: 'targeting_rule'
-    id: string
-    attributes: {
-        description: string
-        key: string
-        name: string
-        operator: Operator
-        ruleVariations: {
-            variationKey: string
-            weight: number
-        }[]
-        segmentKey?: string
-        tags: string[]
-        traitKey?: string
-        traitValue?: string
-        type: 'trait'
-    }
-}
-
-export const createTargetingRule = async (
-    { workspaceKey, projectKey, environmentKey, flagKey }: Partial<FlagbaseParams>,
-    request: TargetingRuleRequest
-): Promise<{ data: TargetingRuleResponse }> => {
-    return axios.post(`/targeting/${workspaceKey}/${projectKey}/${environmentKey}/${flagKey}/rules`, request)
-}
-
-export const deleteTargetingRule = async (
-    { workspaceKey, projectKey, environmentKey, flagKey, ruleKey }: Partial<FlagbaseParams>
-): Promise<{ data: TargetingRuleResponse }> => {
-    return axios.delete(`/targeting/${workspaceKey}/${projectKey}/${environmentKey}/${flagKey}/rules/${ruleKey}`)
-}
