+++ ui/package.json
@@ -108,6 +108,7 @@
         "react-dom": "^18.2.0",
         "react-query": "^3.39.3",
         "react-router-dom": "^6.8.2",
+        "rfc6902": "^5.0.1",
         "tsconfig-paths": "^4.1.0",
         "yup": "^0.32.11"
     },
